feat: define plugin use value dependencies
