+++ crates/rspack_core/src/lib.rs
@@ -3,10 +3,12 @@ mod artifacts;
 mod binding;
 mod compilation;
 mod exports;
+mod value_cache_versions;
 pub use artifacts::*;
 pub use binding::*;
 pub use compilation::make::{ExecuteModuleId, ForwardId, LazyUntil};
 pub use exports::*;
+pub use value_cache_versions::ValueCacheVersions;
 mod dependencies_block;
 pub mod diagnostics;
 pub mod incremental;
