+++ crates/rspack_plugin_dll/src/dll_reference/delegated_module.rs
@@ -8,7 +8,8 @@ use rspack_core::{
   CodeGenerationResult, Compilation, ConcatenationScope, Context, DependenciesBlock, DependencyId,
   FactoryMeta, LibIdentOptions, Module, ModuleDependency, ModuleGraph, ModuleId, ModuleType,
   RuntimeGlobals, RuntimeSpec, SourceType, StaticExportsDependency, StaticExportsSpec,
-  impl_module_meta_info, impl_source_map_config, module_raw, module_update_hash,
+  ValueCacheVersions, impl_module_meta_info, impl_source_map_config, module_raw,
+  module_update_hash,
   rspack_sources::{BoxSource, OriginalSource, RawStringSource},
   throw_missing_module_error_block,
 };
@@ -189,7 +190,7 @@ impl Module for DelegatedModule {
     Ok(code_generation_result)
   }
 
-  fn need_build(&self) -> bool {
+  fn need_build(&self, _value_cache_versions: &ValueCacheVersions) -> bool {
     false
   }
